MSP-3893: Rebased to dev to solve merge Conflicts
MSP-3893: Added a function to add issues directly to a deltaset if they're new, rather than having to be added to the issuemanager.

MSP-3893: Made the planswizard changes only calculate the issue delta set for submission, without applying them. Also resolved some issues with incorrect issue calculation.

MSP-3893: Cleanup and removal or of an unnecessary constraint check

MSP-3893: Moved the network blocker in front of the planswizard