flow: handleImportLayerMetaCallback(List<LayerMeta> layerMeta) -> ServerCommunication.DoRequest<JObject>(Server.GetMELConfig(), form, handleMELConfigCallback) -> handleMELConfigCallback(JObject melConfig) -> LoadFishingFleets(JObject melConfig) -> LoadInitialFishingValues() -> ServerCommunication.DoRequest<List<FishingObject>>(Server.GetInitialFishingValues(), form, LoadInitialFishingValuesCallback) -> LoadInitialFishingValuesCallback(List<FishingObject> fishing) -> initialFishingValues = new FishingDistributionDelta(fishing)
so the issue with the flow is that while MEL is launched, it will need some time to calculate the initial fishing values, and once it has, it will report this to the api using mel/InitialFishing call. However the client is already requesting this information after calling mel/Config, the initial fishing values might not be available yet resulting into an empty array.
Then later you the method is called GetFishingDistributionAtTime(int timeMonth) with:

FishingDistributionSet result = new FishingDistributionSet(initialFishingValues);
result.ApplyValues(plan.fishingDistributionDelta);

with ApplyValues(FishingDistributionDelta deltaSet) having:
    foreach (KeyValuePair<string, Dictionary<int, float>> values in deltaSet.GetValuesByFleet())
    {
        Dictionary<int, float> target = fishingValues[values.Key];

and that will eventually crash becuase it is trying to access the initial finish values using the values.Key which is not there since fishingValues is empty.

So.... to handle this, I just accept the deltaSet values as being the initial values. This is handled by FixMissingInitialFishingValues